backend/requirements.txt
asgiref==3.9.1
bleach==6.2.0
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
